requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
generics==2.0.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
whitenoise==6.0.0
